EPP-160 moved behaviour into common
